[unix] Add endianness conversion to message channels and use architecture-independent uint64_t instead of size_t. Some lint cleanups, too.

+++ src/platforms/unix/stream_sockets_channel.h
@@ -62,7 +62,7 @@ class StreamSocketsChannel : public MessagingChannelInterface<T>,
                         Envelope<T>* final_envelope,
                         typename AsyncRecvHandler<T>::type final_callback);
   void RecvAThirdStage(const boost::system::error_code& error,
-                       const size_t bytes_read, size_t message_size,
+                       const size_t bytes_read, uint64_t message_size,
                        Envelope<T>* final_envelope,
                        typename AsyncRecvHandler<T>::type final_callback);
